first endpoint + rework architecture
