+++ .eslintrc.yml
@@ -1,0 +1,8 @@
+env:
+  browser: true
+  es2021: true
+extends: google
+parserOptions:
+  ecmaVersion: latest
+  sourceType: module
+rules: {}
