+++ src/routes/stmRoutes.js
@@ -1,0 +1,8 @@
+import {Router} from 'express';
+import {getVehiclePosition} from '../controller/stmController.js';
+
+const router = new Router();
+
+router.get('/vehiclePosition', getVehiclePosition);
+
+export default router;
